feat: added "advance" button to TodoCard, unified callback types

+++ frontend/src/TodoCard.tsx
@@ -1,18 +1,23 @@
-import {Todo} from "./Types.tsx";
+import {IdCallback, Todo} from "./Types.tsx";
 import {useNavigate} from "react-router-dom";
 
-export type DeleteTodoCallback = (id: string) => void;
 type TodoCardProps = {
     todo: Todo
-    deleteTodo: DeleteTodoCallback
+    deleteTodo?: IdCallback
+    advanceTodo?: IdCallback
 }
 
-export function TodoCard( props: TodoCardProps ) {
+export default function TodoCard( props: TodoCardProps ) {
     const navigate = useNavigate();
 
     function deleteEntry() {
-        if (props.todo.id)
+        if (props.deleteTodo && props.todo.id)
             props.deleteTodo(props.todo.id)
+    }
+
+    function advanceEntry() {
+        if (props.advanceTodo && props.todo.id)
+            props.advanceTodo(props.todo.id)
     }
 
     return (
@@ -21,7 +26,8 @@ export function TodoCard( props: TodoCardProps ) {
             <div>{props.todo.description}</div>
             <div>{props.todo.status}</div>
             <button onClick={() => navigate("/edit/"+props.todo.id)}>edit</button>
-            <button onClick={deleteEntry}>delete</button>
+            {props. deleteTodo && <button onClick={ deleteEntry}>delete</button>}
+            {props.advanceTodo && <button onClick={advanceEntry}>advance</button>}
         </div>
     )
 }
